Added a popup window after the first completion to give a hint about the auto-continue and disable popup option
Once the user has finished the first cycle and wants to continue, they
probably do not want to read the story again. For a smoother game
experience a popup windiw is shown to give a hint about the options to
auto-continue after completion and the option to prevent story popups.