Added option to split combinator signal output into two: required item and stage signal
The combinator always had the output signal for the required item and a
special signal for the current stage (S). However, sometimes for
automation purposes it is better to have an output only of the required
item.
With this change it is possible to use two combinators. One that outputs
the required item only and the new combinator which outputs the special
signal S for the current stage.